config: Move desiredHostFrequency out of FireSim lib
